メモ追加処理をInitializeNewMentorMemoに分離
